chime4 updates

+++ egs/chime4/mc_asr1/local/real_noisy_chime4_data_prep.sh
@@ -18,7 +18,7 @@ eval_flag=true # make it true when the evaluation data are released
 
 . utils/parse_options.sh || exit 1;
 
-if [ $# -ne 1 ]; then
+if [ $# -ne 3 ]; then
   printf "\nUSAGE: %s <corpus-directory>\n\n" `basename $0`
   echo "The argument should be a the top-level Chime4 directory."
   echo "It is assumed that there will be a 'data' subdirectory"
@@ -28,34 +28,40 @@ fi
 
 echo "$0 $@"  # Print the command line for logging
 
-audio_dir=$1/data/audio/16kHz/isolated
+audio_dir=$1/data/audio/16kHz/$2
 trans_dir=$1/data/transcriptions
+tag=$3
 
-echo "extract all channels (CH1-6.wav) for noisy data"
+echo "extract ${tag} channels (CH*.wav) for noisy data"
 
 dir=`pwd`/data/local/data
 local=`pwd`/local
 utils=`pwd`/utils
 
 if ${eval_flag}; then
-list_set="tr05_real_noisy dt05_real_noisy et05_real_noisy"
+list_set="tr05_real_noisy_${tag} dt05_real_noisy_${tag} et05_real_noisy_${tag}"
 else
-list_set="tr05_real_noisy dt05_real_noisy"
+list_set="tr05_real_noisy_${tag} dt05_real_noisy_${tag}"
 fi
 
 cd ${dir}
 
-find ${audio_dir} -name '*CH[1-6].wav' | grep 'tr05_bus_real\|tr05_caf_real\|tr05_ped_real\|tr05_str_real' | sort -u > tr05_real_noisy.flist
-find ${audio_dir} -name '*CH[1-6].wav' | grep 'dt05_bus_real\|dt05_caf_real\|dt05_ped_real\|dt05_str_real' | sort -u > dt05_real_noisy.flist
+find ${audio_dir} -name '*CH[1-6].wav' | grep 'tr05_bus_real\|tr05_caf_real\|tr05_ped_real\|tr05_str_real' | sort -u > tr05_real_noisy_${tag}.flist
+exit 0
+find ${audio_dir} -name '*CH[1-6].wav' | grep 'dt05_bus_real\|dt05_caf_real\|dt05_ped_real\|dt05_str_real' | sort -u > dt05_real_noisy_${tag}.flist
 if ${eval_flag}; then
-find ${audio_dir} -name '*CH[1-6].wav' | grep 'et05_bus_real\|et05_caf_real\|et05_ped_real\|et05_str_real' | sort -u > et05_real_noisy.flist
+find ${audio_dir} -name '*CH[1-6].wav' | grep 'et05_bus_real\|et05_caf_real\|et05_ped_real\|et05_str_real' | sort -u > et05_real_noisy_${tag}.flist
 fi
 
 # make a dot format from json annotation files
+if ! [ -f dt05_real.dot ]; then
 cp ${trans_dir}/tr05_real.dot_all tr05_real.dot
 cp ${trans_dir}/dt05_real.dot_all dt05_real.dot
+fi
 if ${eval_flag}; then
+if ! [ -f et05_real.dot ]; then
 cp ${trans_dir}/et05_real.dot_all et05_real.dot
+fi
 fi
 
 # make a scp file from file list
@@ -69,18 +75,18 @@ cat tr05_real.dot | sed -e 's/(\(.*\))/\1/' | awk '{print $NF ".CH1_REAL"}'> tr0
 cat tr05_real.dot | sed -e 's/(.*)//' > tr05_real_noisy.txt
 paste -d" " tr05_real_noisy.ids tr05_real_noisy.txt | \
 awk '{print}{sub(/CH1/, "CH2",$0);print}{sub(/CH2/, "CH3",$0);print}{sub(/CH3/, "CH4",$0);print}{sub(/CH4/, "CH5",$0);print}{sub(/CH5/, "CH6",$0);print}' | \
-sort -k 1 > tr05_real_noisy.trans1
+sort -k 1 > tr05_real_noisy_${tag}.trans1
 cat dt05_real.dot | sed -e 's/(\(.*\))/\1/' | awk '{print $NF ".CH1_REAL"}'> dt05_real_noisy.ids
 cat dt05_real.dot | sed -e 's/(.*)//' > dt05_real_noisy.txt
 paste -d" " dt05_real_noisy.ids dt05_real_noisy.txt | \
 awk '{print}{sub(/CH1/, "CH2",$0);print}{sub(/CH2/, "CH3",$0);print}{sub(/CH3/, "CH4",$0);print}{sub(/CH4/, "CH5",$0);print}{sub(/CH5/, "CH6",$0);print}' | \
-sort -k 1 > dt05_real_noisy.trans1
+sort -k 1 > dt05_real_noisy_${tag}.trans1
 if ${eval_flag}; then
 cat et05_real.dot | sed -e 's/(\(.*\))/\1/' | awk '{print $NF ".CH1_REAL"}'> et05_real_noisy.ids
 cat et05_real.dot | sed -e 's/(.*)//' > et05_real_noisy.txt
 paste -d" " et05_real_noisy.ids et05_real_noisy.txt | \
 awk '{print}{sub(/CH1/, "CH2",$0);print}{sub(/CH2/, "CH3",$0);print}{sub(/CH3/, "CH4",$0);print}{sub(/CH4/, "CH5",$0);print}{sub(/CH5/, "CH6",$0);print}' | \
-sort -k 1 > et05_real_noisy.trans1
+sort -k 1 > et05_real_noisy_${tag}.trans1
 fi
 
 # Do some basic normalization steps.  At this point we don't remove OOVs--
@@ -107,6 +113,8 @@ for x in ${list_set}; do
   cp ${x}.txt     ../../${x}/text    || exit 1;
   cp ${x}.spk2utt ../../${x}/spk2utt || exit 1;
   cp ${x}.utt2spk ../../${x}/utt2spk || exit 1;
+  # Check that data dirs are okay!
+  ${utils}/validate_data_dir.sh --no-feats ../../${x} || exit 1
 done
 
 echo "Data preparation succeeded"
